Change Responsability of the gradiant generator
