+++ js/ColorControl.js
@@ -128,44 +128,6 @@ class colorControl {
         }
     }
 
-    static createGradiant(colors, size) {
-        if (colors.length < 2) return;
-
-        let response = [];
-        for (let i = 0; i < colors.length - 1; i++) {
-            const color1 = colors[i];
-            const color2 = colors[i + 1];
-
-            let cantSubColors = size * (color2.position - color1.position) / 100
-
-            if (cantSubColors > 0) {
-                if (i === 0) {
-                    response.push(color1.color);
-                }
-
-                let color1RGB = colorControl.hexToRGb(color1.color);
-                let color2RGB = colorControl.hexToRGb(color2.color);
-
-                let deltaRGB = {
-                    red: parseInt((color2RGB.red - color1RGB.red) / cantSubColors),
-                    green: parseInt((color2RGB.green - color1RGB.green) / cantSubColors),
-                    blue: parseInt((color2RGB.blue - color1RGB.blue) / cantSubColors)
-                };
-
-                for (let sc = 1; sc <= cantSubColors; sc++) {
-                    let nextColor = {
-                        red: color1RGB.red + sc * deltaRGB.red,
-                        green: color1RGB.green + sc * deltaRGB.green,
-                        blue: color1RGB.blue + sc * deltaRGB.blue
-                    };
-                    response.push(colorControl.rgbToHex(nextColor));
-                }
-                response.push(color2.color);
-            }
-        }
-        return response
-    }
-
     static hexToRGb(hex) {
         function cutHex(h) {
             return h.charAt(0) == "#" ? h.substring(1, 7) : h;
